SBERDOMA-128 Before new auth
