+++ apps/condo/public/resetEmail.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html >
+    <head>
+        <meta name="viewport" content="width=device-width" />
+        <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+        <title>Восстановление пароля</title>
+        <style type="text/css">
+            @import url(http://fonts.googleapis.com/css?family=Roboto);
+        </style>
+</head>
+<body style='padding:20px;'>
+    <div style="display: flex; justify-content: space-between; margin-bottom: 20px">
+        <div  style="font-family: Roboto, Arial, sans-serif; color: #000000; font-size: 24px;line-height: 32px; font-weight: 700; margin-right: 20px;">
+            Doma.ai
+        </div>
+        <div style="letter-spacing: 0.3em; font-family: Roboto, Arial, sans-serif; color: #000000; font-size: 16px;line-height: 32px; font-weight: 700;">
+            ВОССТАНОВЛЕНИЕ ПАРОЛЯ
+        </div>
+    </div>
+    <div style="text-align: center; margin-top:60px; margin-bottom:20px; text-align: center; width: 100%;">
+       <img  style='height: 209px; margin: auto;' src='./support.png' ></div>
+    </div>
+
+    <div style="font-family: Roboto, Arial, sans-serif; color: #000000; font-size: 22px;line-height: 32px; max-width: 750px;">
+        Добрый день!<br/>
+        Чтобы задать новый пароль к платформе Doma.ai, вам просто нужно нажать на кнопку ниже.
+    </div>
+    <div style="margin-top:60px;">
+        <a href="{{resetLink}}" style="text-decoration: none; position: relative; display: inline-block; white-space: nowrap; text-align: center; background-image: none; cursor: pointer; transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1); user-select: none; touch-action: manipulation; height: 32px; padding: 4px 15px; border-radius: 4px; background: #fff;background-color:rgb(76, 209, 116); box-shadow: 0 3px 6px #73de90, 0 6px 16px #a0ebb1, 0 9px 28px #d0f7d8; border: 1px solid rgb(76, 209, 116); color: white; font-family: Roboto, Arial, sans-serif; font-size: 16px;line-height: 32px; font-weight: 700;">Задать новый пароль</a>
+    </div>
+</body>
+</html>
+
