+++ apps/condo/schema/User.js
@@ -1,55 +1,16 @@
-const { DateTimeUtc, Relationship } = require('@keystonejs/fields')
+const faker = require('faker') 
 
 const access = require('@core/keystone/access')
 const { ForgotPasswordService } = require('@core/keystone/schemas/User')
-const { ForgotPasswordAction: BaseForgotPasswordAction } = require('@core/keystone/schemas/User')
+const { User: BaseUser } = require('@core/keystone/schemas/User')
+const { RegisterNewUserService } = require('@core/keystone/schemas/User')
+const { Json } = require('@core/keystone/fields')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
 
-const USER_OWNED_FIELD = {
-    schemaDoc: 'Ref to the user. The object will be deleted if the user ceases to exist',
-    type: Relationship,
-    ref: 'User',
-    isRequired: true,
-    knexOptions: { isNotNullable: true }, // Relationship only!
-    kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
-    access: {
-        read: true,
-        create: false, // TODO(pahaz): check access!
-        update: access.userIsAdmin,
-        delete: false,
-    },
-}
+const { admin } = require('@condo/domains/common/utils/firebase.back.utils')
 
-const ForgotPasswordAction = BaseForgotPasswordAction._override({
-    fields: {
-        user: USER_OWNED_FIELD,
 
-        requestedAt: {
-            factory: () => new Date(Date.now()).toISOString(),
-            type: DateTimeUtc,
-            isRequired: true,
-        },
-        expiresAt: {
-            factory: () => new Date(Date.now() + 1000 * 60 * 60 * 24).toISOString(),
-            type: DateTimeUtc,
-            isRequired: true,
-        },
-        usedAt: {
-            type: DateTimeUtc,
-            defaultValue: null,
-        },
-    },
-    plugins: [uuided(), versioned(), tracked(), softDeleted(), historical()],
-    access: {
-        auth: true,
-        read: access.userIsAdmin,
-        create: access.userIsAdmin,
-        update: access.userIsAdmin,
-        delete: access.userIsAdmin,
-    },
-})
 
 module.exports = {
-    ForgotPasswordAction,
-    ForgotPasswordService,
+    RegisterNewUserService,
 }
